% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 3.R
\name{get_correlation_matrix}
\alias{get_correlation_matrix}
\title{Limit data to numeric continuous variables for a given year}
\usage{
get_correlation_matrix(year)
}
\arguments{
\item{year}{The year for which the data is to be retrieved.}
}
\value{
A correlation matrix for all numeric continuous variables in the data for the given year.
}
\description{
This function takes an individual year as input and returns the correlation matrix for all numeric continuous variables in the dataset for that year.
}
\examples{
correlation_matrix <- get_correlation_matrix(2015)

}
