% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 2.R
\name{get_oldest_player}
\alias{get_oldest_player}
\title{Get the oldest player in the NBA dataset for a given year}
\usage{
get_oldest_player(year)
}
\arguments{
\item{year}{The year for which the oldest player is to be retrieved.}
}
\value{
A data frame containing the data for the oldest player in the given year.
}
\description{
This function takes an individual year as input and returns the data for the oldest player in the NBA dataset for that year. If multiple players have the same age, the function will return the first player in the dataset with that age.
}
\examples{
get_oldest_player(2012)

}
