% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 1.R
\name{get_player_data}
\alias{get_player_data}
\title{NBA Function}
\usage{
get_player_data(year)
}
\arguments{
\item{love}{}
}
\description{
This function allows you to get data point of interest for a player in a given year
}
\examples{
get_player_data()
}
